backend/pyproject.toml
[project]
name = "rizzbot-backend"
version = "2.0.0"
description = "RizzBot V2 — AI dating chat reply engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "slowapi>=0.1.9",
    "structlog>=24.0",
    "sentry-sdk[fastapi]>=2.0",
    "firebase-admin>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]
postgres = [
    "asyncpg>=0.30",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
